LibWeb: Move the computed style accessor surface into the css module
The property accessors on StyleValues — which field of which style
group carries width, the margins, tab-size and the rest — are
computed-values knowledge, the Rust twin of the accessor layer in the
C++ ComputedValues.h, so they now live beside the payload types as
ComputedValuesView in computed_value_views.rs. The view owns the group
pointer array, the typed group getters and every pure read, including
the display facts and the style-only block-formatting-context
predicate, with the group indices moving into computed_value_types.rs
where the C++ static asserts keep pinning them through the same
generated header.

StyleValues shrinks to the layout pass state the css module cannot
know about — the anchor-inset store behind the four inset reads and
the line builder's vertical-align keyword substitution — and derefs to
the view for everything else, so call sites are unchanged. This also
undoes the optional-store compromise from folding StyleReader away:
contexts outside a pass (tree building, node facts) now hold a plain
ComputedValuesView with no inset surface at all, and the pass view
carries the store unconditionally again.
