+++ Libraries/LibWeb/Rust/src/css/computed_value_types.rs
@@ -369,3 +369,18 @@ pub struct SVGResetValues {
     pub vector_effect: u8,
     pub shape_rendering: u8,
 }
+
+// Registered indices of the style groups the computed-values view reads,
+// pinned to the C++ StyleGroupIndex enum by static_asserts in
+// LayoutRustBridge.cpp.
+pub const STYLE_GROUP_INDEX_INHERITED_TABLE: usize = 0;
+pub const STYLE_GROUP_INDEX_GRID: usize = 9;
+pub const STYLE_GROUP_INDEX_INHERITED_TEXT: usize = 4;
+pub const STYLE_GROUP_INDEX_INHERITED_BOX: usize = 5;
+pub const STYLE_GROUP_INDEX_FONT: usize = 6;
+pub const STYLE_GROUP_INDEX_SVG_RESET: usize = 8;
+pub const STYLE_GROUP_INDEX_BORDER: usize = 17;
+pub const STYLE_GROUP_INDEX_ALIGNMENT: usize = 18;
+pub const STYLE_GROUP_INDEX_SIZING: usize = 20;
+pub const STYLE_GROUP_INDEX_SURROUND: usize = 21;
+pub const STYLE_GROUP_INDEX_BOX: usize = 22;
