+++ Libraries/LibWeb/Rust/src/layout/table_formatting_context.rs
@@ -808,7 +808,7 @@ impl TableTree for TableFormattingContext<'_> {
 
     fn display(&self, node: Node) -> FfiDisplay {
         self.callbacks
-            .style_values_if_styled(node)
+            .computed_values_view_if_styled(node)
             .map_or_else(FfiDisplay::block, |style| style.display())
     }
